+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.0 (June 02, 2020)
+  - Merge branch 'feature/base-code' into develop
+  - updated requirements.txt file
+  - update docs
+  - update docs
+  - new apis: auth login, logout, get login status
+  - added sample insert data into table `users`
+  - added default codes: Flask App + Swagger UI
+
 ## 0.1.0 (May 18, 2020)
   - added default files
   - Initial commit